% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{content.from.endpoint}
\alias{content.from.endpoint}
\title{Content from endpoint}
\usage{
content.from.endpoint(
  url,
  content_type = "application/octet-stream",
  accept = "application/json",
  fromJSON = TRUE
)
}
\arguments{
\item{url}{API URL}

\item{content_type}{(default "application/octet-stream")}

\item{accept}{(default "application/json")}

\item{fromJSON}{(default TRUE)}
}
\value{
encoded
}
\description{
Content from endpoint
}
\examples{
content.from.endpoint("https://api-web.nhle.com/v1/standings-season")
}
