% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{count.decimals}
\alias{count.decimals}
\title{count.decimals}
\usage{
count.decimals(number)
}
\arguments{
\item{number}{number}
}
\value{
integer
}
\description{
Counts the number of decimal places in provided number.
}
\examples{
count.decimals(0.00987)
}
