% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{count.n}
\alias{count.n}
\title{count.n}
\usage{
count.n(frame, columns = NA)
}
\arguments{
\item{frame}{frame}

\item{columns}{one or more columns (default is NA)}
}
\value{
integer
}
\description{
Counts the number of unique elements in specified columns of a provided frame or the frame as a whole.
}
\examples{
frame <- data.frame("a"=rep(1, 9), "b"=c(rep(1, 3), rep(2, 3), rep(3, 3)), "c"=1:9)
count.n(frame)
count.n(frame, c("a", "b"))
}
