% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{curve.equation}
\alias{curve.equation}
\title{Curve equation}
\usage{
curve.equation(x.y, x1, y1, x2, y2, a)
}
\arguments{
\item{x.y}{either "x" or "y"}

\item{x1}{x coord for first point}

\item{y1}{y coord for first point}

\item{x2}{x coord for second point}

\item{y2}{y coord for second point}

\item{a}{a value of curve}
}
\value{
list(pos.function, neg.function)
}
\description{
Curve equation
}
\examples{
curve.equation("y", 0, 0, 3, 3, 0.7)
}
