% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{duplicated.values}
\alias{duplicated.values}
\title{Duplicated values}
\usage{
\method{duplicated}{values}(vector, non.duplicated = FALSE)
}
\arguments{
\item{vector}{vector}

\item{non.duplicated}{boolean switch between duplicated and non duplicated values (default is FALSE)}
}
\value{
vector
}
\description{
Identifies all of the duplicated values in a vector.
}
\examples{
frame <- data.frame("a"=rep(1, 9), "b"=c(rep(1, 3), rep(2, 3), rep(3, 3)), "c"=1:9)
duplicated.values(frame[,"b"])
}
