% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{linear.equation}
\alias{linear.equation}
\title{Linear equation}
\usage{
linear.equation(x = NA, y = NA, slope = NA, equation = NA)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}

\item{slope}{numeric}

\item{equation}{named list("slope"=numeric, "b"=numeric)}
}
\value{
varies
}
\description{
Creates a linear equation with the given information.
Can provide x, y, and slope or y and equation
linear.equation(y, equation) out:list(x, y)
linear.equation(x, equation) out:
list(x, y)
linear.equation(x, y, slope) out:list(slope, b)
}
\examples{
linear.equation(3, 9, 4/5)
}
