% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{locate.element}
\alias{locate.element}
\title{locate.element}
\usage{
locate.element(frame, value, column)
}
\arguments{
\item{frame}{frame}

\item{value}{value to be identified}

\item{column}{column}
}
\value{
vector
}
\description{
Identifies location of specified element within a column of the frame.
}
\examples{
frame <- data.frame("a"=rep(1, 9), "b"=c(rep(1, 3), rep(2, 3), rep(3, 3)), "c"=1:9)
locate.element(frame, 1, "b")
}
