% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{merge.by.overlap}
\alias{merge.by.overlap}
\title{merge.by.overlap}
\usage{
\method{merge}{by.overlap}(frame.list, all = FALSE)
}
\arguments{
\item{frame.list}{list of frames}

\item{all}{boolean switch determining inclusion of NA cells (default is FALSE)}
}
\value{
dataframe
}
\description{
Merges two data frames by their overlapping columns.
}
\examples{
frame <- data.frame("a"=rep(1, 9), "b"=c(rep(1, 3), rep(2, 3), rep(3, 3)), "c"=1:9)
frame_2 <- data.frame(frame, "d"=rowSums(frame))
merge.by.overlap(list(frame, frame_2))
}
