% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.replace.R
\name{remove.na}
\alias{remove.na}
\title{remove.na}
\usage{
remove.na(object, column = NA)
}
\arguments{
\item{object}{frame, list, or vector}

\item{column}{column with NA if object is a frame (default is NA)}
}
\value{
object
}
\description{
Removes all rows with NA values for an object (frame, list, or vector).
}
\examples{
remove.na(list(NA, 1, NA, 2, 3))
}
