% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{rename.column}
\alias{rename.column}
\title{rename.column}
\usage{
rename.column(frame, colname.from, colname.to)
}
\arguments{
\item{frame}{frame}

\item{colname.from}{existing column name}

\item{colname.to}{relacement column name}
}
\value{
frame
}
\description{
(relys on which.col.name) Renames specified column in frame.
}
\examples{
frame <- data.frame("a"=rep(1, 9), "b"=c(rep(1, 3), rep(2, 3), rep(3, 3)), "c"=1:9)
rename.column(frame, "c", "d")
}
