% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.replace.R
\name{replace.na}
\alias{replace.na}
\title{replace.na}
\usage{
replace.na(object, value)
}
\arguments{
\item{object}{frame, list, or vector}

\item{value}{value to replace NA}
}
\value{
object
}
\description{
Replaces all NA in object with specified value
}
\examples{
replace.na(list(NA, 1, NA, 2, 3), 0)
}
