% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.replace.R
\name{replace.nan}
\alias{replace.nan}
\title{replace.nan}
\usage{
replace.nan(object, value)
}
\arguments{
\item{object}{frame, list, or vector}

\item{value}{value to replace NaN}
}
\value{
object
}
\description{
Replaces all NaN in object with specified value
}
\examples{
replace.nan(c(NaN, 1, NaN, 2, 3), 0)
}
