% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.replace.R
\name{replace.specific}
\alias{replace.specific}
\title{replace.specific}
\usage{
replace.specific(vector, value, replacement)
}
\arguments{
\item{vector}{vector}

\item{value}{existing value}

\item{replacement}{replacement value}
}
\value{
vector
}
\description{
Locates all instances of a value and replaces them with the provided value.
}
\examples{
replace.specific(c(1, 2, 3), 1, 3)
}
