% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{round.to}
\alias{round.to}
\title{round.to}
\usage{
\method{round}{to}(number, to, top.bottom = NA)
}
\arguments{
\item{number}{number}

\item{to}{decimal place for rounding (accepts 0.001 or alike)}

\item{top.bottom}{switch to round up or down (default is traditional rounding)}
}
\value{
numeric
}
\description{
Round the number to a the provided decimal place (0.001)
}
\examples{
round.to(0.00987, 0.001, "bottom")
}
