% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{row.bind}
\alias{row.bind}
\title{row.bind}
\usage{
row.bind(frames)
}
\arguments{
\item{frames}{list of data frames}
}
\value{
frame
}
\description{
Binds multiple frames together by appending rows.
}
\examples{
frame <- data.frame("a"=rep(1, 9), "b"=c(rep(1, 3), rep(2, 3), rep(3, 3)), "c"=1:9)
row.bind(list(frame, frame))
}
