% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{row.names.to.column}
\alias{row.names.to.column}
\title{row.names.to.column}
\usage{
\method{row.names}{to.column}(frame, numerical.rows = FALSE)
}
\arguments{
\item{frame}{data frame}

\item{numerical.rows}{boolean switch specifying if row names should be numeric (default is FALSE)}
}
\value{
frame
}
\description{
Preserves a frames row names as a column.
}
\examples{
frame <- data.frame("a"=rep(1, 9), "b"=c(rep(1, 3), rep(2, 3), rep(3, 3)), "c"=1:9)
rownames(frame) <- LETTERS[1:9]
row.names.to.column(frame)
}
