% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{sigfill}
\alias{sigfill}
\title{sigfill}
\usage{
sigfill(numbers, sigfigs = 2, as.p = FALSE)
}
\arguments{
\item{numbers}{numbers}

\item{sigfigs}{number of significant figures}

\item{as.p}{boolean switch between p-value formating (default is FALSE)}
}
\value{
strings
}
\description{
Rounds number to desired significance and formats as a p-value.
}
\examples{
sigfill(c(0.00123, 0.000987), 3, TRUE)
}
