% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{slope}
\alias{slope}
\title{Slope}
\usage{
slope(point.1, point.2)
}
\arguments{
\item{point.1}{coords of first point c(x, y)}

\item{point.2}{coords of second point c(x, y)}
}
\value{
slope
}
\description{
Slope
}
\examples{
slope (c(0,3), c(9, 6))
}
