% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{subset.object}
\alias{subset.object}
\title{subset.object}
\usage{
\method{subset}{object}(object, value, col.name = NULL, remove = FALSE)
}
\arguments{
\item{object}{frame, list, or vector}

\item{value}{value}

\item{col.name}{column with value}

\item{remove}{boolean switch between retaining only the specified value or remove the value from the object (default is FALSE)}
}
\value{
object
}
\description{
Subsets or removes specified values provided object (frame, list, or vector).
}
\examples{
frame <- data.frame("a"=rep(1, 9), "b"=c(rep(1, 3), rep(2, 3), rep(3, 3)), "c"=1:9)
subset.object(frame, 1, "b")
}
