% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{two.by.two.paired}
\alias{two.by.two.paired}
\title{two.by.two.paired}
\usage{
two.by.two.paired(frame, col.1, col.2)
}
\arguments{
\item{frame}{frame}

\item{col.1}{column of numeric values (0 or value > 0).}

\item{col.2}{column of numeric values (0 or value > 0).}
}
\value{
table
}
\description{
Compares two columns in a frame to generate a two by two table summarizing presence of a value (>0).
}
\examples{
frame <- data.frame("a"=c(3, 0.12, 0, 0, 0.001), "b"=c(0, 1, 0, 6, 9))
two.by.two.paired(frame, "a", "b")
}
