% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{uln}
\alias{uln}
\title{Unlist unname}
\usage{
uln(object)
}
\arguments{
\item{object}{list or vector}
}
\value{
object
}
\description{
Unlists and unnames the provided object.
}
\examples{
frame <- data.frame("a"=rep(1, 9), "b"=c(rep(1, 3), rep(2, 3), rep(3, 3)), "c"=1:9)
uln(frame[,"b"])
}
