% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{uuln}
\alias{uuln}
\title{Unlist unname unique}
\usage{
uuln(object)
}
\arguments{
\item{object}{list or vector}
}
\value{
object
}
\description{
Unlists, unnames, and identifies the unique values in the provided object (list or vector).
}
\examples{
frame <- data.frame("a"=rep(1, 9), "b"=c(rep(1, 3), rep(2, 3), rep(3, 3)), "c"=1:9)
uuln(frame[,"b"])
}
