% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{which.col.name}
\alias{which.col.name}
\title{which.col.name}
\usage{
which.col.name(dataframe, col.name)
}
\arguments{
\item{dataframe}{data frame}

\item{col.name}{column}
}
\value{
vector
}
\description{
(replace with which col row name) Identifies location of specified columns in frame.
}
\examples{
which.col.name(frame, c("c", "a", "b"))
}
